Added a TypeResolver (found here: http://blogs.msdn.com/b/youssefm/archive/2009/06/05/introducing-a-new-datacontractserializer-feature-the-datacontractresolver.aspx)
Fixed the Base64Serialize class to close the writer and reader, not just te stream. This causes it to flush and save the file correctly.

There are currently problems with KnownTypes, possible fix will be to move serialization stuff into MayhemCore and register known types in ModuleList as they are loaded.