Modify the build batch file to get the packages from the new location.
Added nuspecs for Mayhem itself and the Updater application. Currently the build file copies all of the files into a lib folder. Since this is unintended behavior, until a new build file is written, these packages will have to be modified by hand before release to be in the correct structure.